수고하셨습니다

+++ src/main/java/com/codeQuartette/myTime/service/impl/HabitServiceImpl.java
@@ -7,13 +7,17 @@ import com.codeQuartette.myTime.domain.HabitHasMyDate;
 import com.codeQuartette.myTime.domain.MyDate;
 import com.codeQuartette.myTime.domain.User;
 import com.codeQuartette.myTime.domain.value.Category;
+import com.codeQuartette.myTime.exception.CannotCreateHabitException;
 import com.codeQuartette.myTime.exception.HabitNotFoundException;
+import com.codeQuartette.myTime.exception.UserNotFoundException;
+import com.codeQuartette.myTime.exception.UserNotMatchException;
 import com.codeQuartette.myTime.repository.HabitRepository;
 import com.codeQuartette.myTime.service.HabitHasMyDateService;
 import com.codeQuartette.myTime.service.HabitService;
 import com.codeQuartette.myTime.service.MyDateService;
 import com.codeQuartette.myTime.service.UserService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.cache.annotation.CacheEvict;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -34,12 +38,16 @@ public class HabitServiceImpl implements HabitService {
 
     @Override
     @Transactional
+    @CacheEvict(cacheNames = "find-habit", allEntries = true)
     public void create(Long userId, HabitDTO.Request habitRequestDTO) {
         User user = userService.findUser(userId);
         Habit habit = Habit.create(habitRequestDTO);
         List<LocalDate> allHabitDates =
                 myDateService.checkAllDateByStartDateAndEndDate(habit.getStartDate(), habit.getEndDate(), habitRequestDTO.getRepeatDay());
         List<MyDate> newMyDates = allHabitDates.stream().map(date -> new MyDate(date, user)).toList();
+        if(newMyDates.size() == 0) {
+            throw new CannotCreateHabitException("습관에 해당하는 날짜가 없습니다. 다시 확인 후 습관을 생성해주세요.");
+        }
         List<MyDate> saveMyDates = myDateService.saveAllMyDate(newMyDates);
 
         habit = habitRepository.save(habit);
@@ -52,6 +60,8 @@ public class HabitServiceImpl implements HabitService {
     }
 
     @Override
+    @Transactional
+    @CacheEvict(cacheNames = "find-habit", allEntries = true)
     public void update(Long userId, Long id, HabitDTO.Request habitRequestDTO) {
         User user = userService.findUser(userId);
         Habit habit = habitRepository.findById(id)
@@ -74,16 +84,27 @@ public class HabitServiceImpl implements HabitService {
     }
 
     @Override
-    public void delete(Long id) {
+    @CacheEvict(cacheNames = "find-habit", allEntries = true)
+    public void delete(Long userId, Long id) {
         Habit habit = habitRepository.findById(id)
                 .orElseThrow(() -> new HabitNotFoundException("삭제하려는 습관을 조회할 수 없습니다."));
+
+        if(!habit.getHabitHasMyDates().get(0).getMyDate().matchUser(userId)) {
+            throw new UserNotMatchException();
+        }
         habitRepository.delete(habit);
     }
 
     @Override
-    public Habit findHabit(Long id) {
-        return habitRepository.findById(id)
+    @Transactional
+    public Habit findHabit(Long userId, Long id) {
+        Habit habit = habitRepository.findById(id)
                 .orElseThrow(() -> new HabitNotFoundException("해당 습관을 조회할 수 없습니다."));
+
+        if(!habit.getHabitHasMyDates().get(0).getMyDate().matchUser(userId)) {
+            throw new UserNotMatchException();
+        }
+        return habit;
     }
 
     @Override
